requirements.txt
torch
torchvision
timm
pillow
numpy
scikit-learn
matplotlib
pandas
